# docs/swagger.yaml
basePath: /
definitions:
...
  domain.UserResponse:
    properties:
      avatarFileId:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      bio:
        example: Backend разработчик
        type: string
      createdAt:
        example: "2026-03-19T13:18:48.000Z"
        format: date-time
        type: string
      displayName:
        example: Иван Иванов
        type: string
      email:
        example: student@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        example: "+79991234567"
        type: string
      updatedAt:
        example: "2026-03-19T13:18:48.000Z"
        format: date-time
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        example: student@example.com
        format: email
        type: string
      password:
        example: StrongPass123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      accessExpiresIn:
        example: 15m
        type: string
      message:
        example: успешный вход
        type: string
      refreshExpiresIn:
        example: 7d
        type: string
    type: object
  dto.LogoutAllResponse:
    properties:
      message:
        example: успешный выход из всех сессий
        type: string
    type: object
  dto.LogoutResponse:
    properties:
      message:
        example: успешный выход
        type: string
    type: object
...
  dto.RefreshResponse:
    properties:
      accessExpiresIn:
        example: 15m
        type: string
      message:
        example: токены обновлены
        type: string
      refreshExpiresIn:
        example: 7d
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: student@example.com
        format: email
        type: string
      password:
        example: StrongPass123
        minLength: 8
        type: string
      phone:
        example: "+79991234567"
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterResponse:
    properties:
      message:
        example: пользователь успешно зарегистрирован
        type: string
      userId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      avatarFileId:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      bio:
        example: Backend разработчик
        type: string
      displayName:
        example: Иван Иванов
        type: string
    type: object
...
  handler.AuthErrorResponse:
    properties:
      error:
        example: invalid_or_expired_token
        type: string
    type: object
...
host: localhost:4200
info:
  contact: {}
  description: 'REST API на Go (Gin + MongoDB + Redis + MinIO + RabbitMQ) с JWT/OAuth2-аутентификацией,
    CRUD-ресурсами, файловым хранилищем, асинхронной обработкой событий и Kubernetes-деплоем.
    ЛР9: health-зонды (/health/live, /health/ready), K8s-манифесты, горизонтальное
    масштабирование, Redis distributed lock.'
  title: Lab 2–9 REST API
  version: "1.4"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: успешный вход (JWT в HttpOnly cookies)
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Вход пользователя
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LogoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      security:
      - CookieAuth: []
      summary: Выход из текущей сессии
      tags:
      - auth
  /auth/logout-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LogoutAllResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      security:
      - CookieAuth: []
      summary: Выход из всех сессий
      tags:
      - auth
...
  /auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: успешное обновление (JWT в HttpOnly cookies)
          schema:
            $ref: '#/definitions/dto.RefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Обновление токенов
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: успешная регистрация
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
...
  /auth/whoami:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: данные текущего пользователя
          schema:
            $ref: '#/definitions/domain.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      security:
      - CookieAuth: []
      summary: Получение данных текущего пользователя
      tags:
      - auth
...
  /profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      security:
      - CookieAuth: []
      summary: Получение профиля текущего пользователя
      tags:
      - profile
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные обновления профиля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      security:
      - CookieAuth: []
      summary: Обновление профиля текущего пользователя
      tags:
      - profile
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"