# internal/auth/handler/auth_handler.go
// Package handler содержит HTTP-обработчики (контроллеры) для модуля авторизации.
// Примечание: Комментарии с аннотациями @Summary, @Tags, @Param и т.д.
// используются для автоматической генерации API-документации через Swagger/OpenAPI.
package handler

import (
	"errors"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"

	"github.com/lab2/rest-api/internal/auth/domain"
	"github.com/lab2/rest-api/internal/auth/dto"
	"github.com/lab2/rest-api/internal/auth/service"
)

// AuthHandler обрабатывает HTTP запросы для авторизации
type AuthHandler struct {
	authService service.AuthService
}

// NewAuthHandler создаёт новый экземпляр хендлера
func NewAuthHandler(authService service.AuthService) *AuthHandler {
	return &AuthHandler{
		authService: authService,
	}
}

// Register обрабатывает POST /auth/register
// @Summary Регистрация пользователя
// @Tags auth
// @Accept json
// @Produce json
// @Param request body dto.RegisterRequest true "Данные для регистрации"
// @Success 201 {object} dto.RegisterResponse "успешная регистрация"
// @Failure 400 {object} AuthErrorResponse
// @Failure 401 {object} AuthErrorResponse
// @Failure 403 {object} AuthErrorResponse
// @Failure 404 {object} AuthErrorResponse
// @Failure 500 {object} AuthErrorResponse
// @Router /auth/register [post]
func (h *AuthHandler) Register(c *gin.Context) {
	var req dto.RegisterRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "некорректные данные запроса"})
		return
	}

	user, err := h.authService.Register(c.Request.Context(), &req)
	if err != nil {
		if errors.Is(err, errors.New("пользователь с таким email уже существует")) {
			c.JSON(http.StatusConflict, gin.H{"error": err.Error()})
			return
		}
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusCreated, dto.RegisterResponse{
		Message: "пользователь успешно зарегистрирован",
		UserID:  user.ID,
	})
}

// Login обрабатывает POST /auth/login
// @Summary Вход пользователя
// @Tags auth
// @Accept json
// @Produce json
// @Param request body dto.LoginRequest true "Данные для входа"
// @Success 200 {object} dto.LoginResponse "успешный вход (JWT в HttpOnly cookies)"
// @Failure 400 {object} AuthErrorResponse
// @Failure 401 {object} AuthErrorResponse
// @Failure 403 {object} AuthErrorResponse
// @Failure 404 {object} AuthErrorResponse
// @Failure 500 {object} AuthErrorResponse
// @Router /auth/login [post]
func (h *AuthHandler) Login(c *gin.Context) {
	var req dto.LoginRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "некорректные данные запроса"})
		return
	}

	tokens, err := h.authService.Login(c.Request.Context(), req.Email, req.Password)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}

	// Установка HttpOnly cookies
	// Access Token (короткое время жизни)
	c.SetCookie(
		"access_token",
		tokens.AccessToken, // Токен устанавливается в authService, здесь только метаданные
		int(tokens.AccessExpiresIn.Seconds()),
		"/",
		"",
		false, // Secure: false для localhost, true для HTTPS
		true,  // HttpOnly: true (JavaScript не имеет доступа)
	)

	// Refresh Token (длительное время жизни)
	c.SetCookie(
		"refresh_token",
		tokens.RefreshToken,
		int(tokens.RefreshExpiresIn.Seconds()),
		"/",
		"",
		false,
		true,
	)

	c.JSON(http.StatusOK, dto.LoginResponse{
		Message:          "успешный вход",
		AccessExpiresIn:  tokens.AccessExpiresIn.String(),
		RefreshExpiresIn: tokens.RefreshExpiresIn.String(),
	})
}

// Refresh обрабатывает POST /auth/refresh
// @Summary Обновление токенов
// @Tags auth
// @Produce json
// @Success 200 {object} dto.RefreshResponse "успешное обновление (JWT в HttpOnly cookies)"
// @Failure 400 {object} AuthErrorResponse
// @Failure 401 {object} AuthErrorResponse
// @Failure 403 {object} AuthErrorResponse
// @Failure 404 {object} AuthErrorResponse
// @Failure 500 {object} AuthErrorResponse
// @Router /auth/refresh [post]
func (h *AuthHandler) Refresh(c *gin.Context) {
	// Получаем refresh токен из cookies
	refreshToken, err := c.Cookie("refresh_token")
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "refresh токен не найден"})
		return
	}

	tokens, err := h.authService.Refresh(c.Request.Context(), refreshToken)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}

	// Обновляем cookies с НОВЫМИ токенами
	// Access Token
	c.SetCookie(
		"access_token",
		tokens.AccessToken, // ← Было: "", стало: tokens.AccessToken
		int(tokens.AccessExpiresIn.Seconds()),
		"/",
		"",
		false,
		true,
	)

	// Refresh Token
	c.SetCookie(
		"refresh_token",
		tokens.RefreshToken, // ← Было: "", стало: tokens.RefreshToken
		int(tokens.RefreshExpiresIn.Seconds()),
		"/",
		"",
		false,
		true,
	)

	c.JSON(http.StatusOK, dto.RefreshResponse{
		Message:          "токены обновлены",
		AccessExpiresIn:  tokens.AccessExpiresIn.String(),
		RefreshExpiresIn: tokens.RefreshExpiresIn.String(),
	})
}

// WhoAmI обрабатывает GET /auth/whoami
// @Summary Получение данных текущего пользователя
// @Tags auth
// @Produce json
// @Security CookieAuth
// @Success 200 {object} domain.UserResponse "данные текущего пользователя"
// @Failure 400 {object} AuthErrorResponse
// @Failure 401 {object} AuthErrorResponse
// @Failure 403 {object} AuthErrorResponse
// @Failure 404 {object} AuthErrorResponse
// @Failure 500 {object} AuthErrorResponse
// @Router /auth/whoami [get]
func (h *AuthHandler) WhoAmI(c *gin.Context) {
	if current, ok := c.Get("currentUser"); ok {
		if user, valid := current.(*domain.User); valid {
			c.JSON(http.StatusOK, user.ToResponse())
			return
		}
	}
	// UserID добавляется в контекст middleware
	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "пользователь не авторизован"})
		return
	}

	userUUID, err := uuid.Parse(userID.(string))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "ошибка обработки запроса"})
		return
	}

	user, err := h.authService.GetUserByID(c.Request.Context(), userUUID)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, user)
}

// Logout обрабатывает POST /auth/logout
// @Summary Выход из текущей сессии
// @Tags auth
// @Produce json
// @Security CookieAuth
// @Success 200 {object} dto.LogoutResponse
// @Failure 400 {object} AuthErrorResponse
// @Failure 401 {object} AuthErrorResponse
// @Failure 403 {object} AuthErrorResponse
// @Failure 404 {object} AuthErrorResponse
// @Failure 500 {object} AuthErrorResponse
// @Router /auth/logout [post]
func (h *AuthHandler) Logout(c *gin.Context) {
	refreshToken, err := c.Cookie("refresh_token")
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "refresh токен не найден"})
		return
	}

	accessToken, _ := c.Cookie("access_token")
	if err := h.authService.Logout(c.Request.Context(), refreshToken, accessToken); err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "ошибка при выходе"})
		return
	}

	// Удаляем cookies
	c.SetCookie("access_token", "", -1, "/", "", false, true)
	c.SetCookie("refresh_token", "", -1, "/", "", false, true)

	c.JSON(http.StatusOK, dto.LogoutResponse{
		Message: "успешный выход",
	})
}

// LogoutAll обрабатывает POST /auth/logout-all
// @Summary Выход из всех сессий
// @Tags auth
// @Produce json
// @Security CookieAuth
// @Success 200 {object} dto.LogoutAllResponse
// @Failure 400 {object} AuthErrorResponse
// @Failure 401 {object} AuthErrorResponse
// @Failure 403 {object} AuthErrorResponse
// @Failure 404 {object} AuthErrorResponse
// @Failure 500 {object} AuthErrorResponse
// @Router /auth/logout-all [post]
func (h *AuthHandler) LogoutAll(c *gin.Context) {
	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "пользователь не авторизован"})
		return
	}

	userUUID, err := uuid.Parse(userID.(string))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "ошибка обработки запроса"})
		return
	}

	if err := h.authService.LogoutAll(c.Request.Context(), userUUID); err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "ошибка при выходе из всех сессий"})
		return
	}

	// Удаляем cookies
	c.SetCookie("access_token", "", -1, "/", "", false, true)
	c.SetCookie("refresh_token", "", -1, "/", "", false, true)

	c.JSON(http.StatusOK, dto.LogoutAllResponse{
		Message: "успешный выход из всех сессий",
	})
}

// GetProfile обрабатывает GET /profile
// @Summary Получение профиля текущего пользователя
// @Tags profile
// @Produce json
// @Security CookieAuth
// @Success 200 {object} domain.UserResponse
// @Failure 401 {object} AuthErrorResponse
// @Failure 404 {object} AuthErrorResponse
// @Failure 500 {object} AuthErrorResponse
// @Router /profile [get]
func (h *AuthHandler) GetProfile(c *gin.Context) {
	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "пользователь не авторизован"})
		return
	}
	userUUID, err := uuid.Parse(userID.(string))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "ошибка обработки запроса"})
		return
	}
	profile, err := h.authService.GetUserByID(c.Request.Context(), userUUID)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, profile)
}

// UpdateProfile обрабатывает POST /profile
// @Summary Обновление профиля текущего пользователя
// @Tags profile
// @Accept json
// @Produce json
// @Security CookieAuth
// @Param request body dto.UpdateProfileRequest true "Данные обновления профиля"
// @Success 200 {object} domain.UserResponse
// @Failure 400 {object} AuthErrorResponse
// @Failure 401 {object} AuthErrorResponse
// @Failure 404 {object} AuthErrorResponse
// @Failure 500 {object} AuthErrorResponse
// @Router /profile [post]
func (h *AuthHandler) UpdateProfile(c *gin.Context) {
	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "пользователь не авторизован"})
		return
	}
	userUUID, err := uuid.Parse(userID.(string))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "ошибка обработки запроса"})
		return
	}

	var req dto.UpdateProfileRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "некорректные данные запроса"})
		return
	}

	profile, err := h.authService.UpdateProfile(c.Request.Context(), userUUID, &req)
	if err != nil {
		if errors.Is(err, service.ErrAvatarOwnership) {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		} else {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "внутренняя ошибка сервера"})
		}
		return
	}
	c.JSON(http.StatusOK, profile)
}